+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `pipeline.pipeline()` by [@natanfeitosa](https://github.com/natanfeitosa/) ([#2](https://github.com/natanfeitosa/pyfunctools/issues/2))
+
 ### Changed
 
 ### Fixed